fawkes-robotino: add explore_zone_wrapped
As we have to call the explore_zone skill out of yagi but provide zone
borders we have to make use of a wrapper skill.
This skill accepts simply a zone-name (e.g. "Z9") and calculates the
particular min-x, min-y, max-x and max-y values required for the
explore_zone skill.
This calculation is taken from the clips agent located in
fawkes-robotino/src/agents/llsf2015/exploration.clp